moved Abstractions into a single file
