--- src/FAI.Core/Abstractions/IInferenceStepsBuilder.cs
@@ -1,6 +1,0 @@
-namespace FAI.Core.Abstractions;
-
-public interface IInferenceStepsBuilder<TInput, TOutput, TInference> where TInference : IInferenceSteps<TInput, TOutput>
-{
-    ValueTask<TInference> BuildAsync();
-}
