--- src/FAI.Core/Abstractions/IPipelineBatchExecutorBuilder.cs
@@ -1,6 +1,0 @@
-namespace FAI.Core.Abstractions;
-
-public interface IPipelineBatchExecutorBuilder<TInput, TOutput>
-{
-    ValueTask<IPipelineBatchExecutor<TInput, TOutput>> BuildAsync();
-}
