--- src/FAI.Core/Abstractions/IPreprocessor.cs
@@ -1,8 +1,0 @@
-using System.Numerics.Tensors;
-
-namespace FAI.Core.Abstractions;
-
-public interface IPreprocessor<TInput, out TPreprocessContainer, TFloat> where TPreprocessContainer : IEnumerable<Tensor<TFloat>>
-{
-    TPreprocessContainer Preprocess(ReadOnlySpan<TInput> input);
-}
